requirements.txt
django
djangorestframework
django-cors-headers
django-extensions
coverage
pylint
